Making test files for each graph
